requirements.txt
phidata
python-dotenv
python-docx
python-multipart
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
openai